requirements.txt
Flask
SQLAlchemy
python-dotenv
pytest
requests
pymysql
